+++ bin/wasm-node/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Block headers with a digest item of type `Other` no longer fail to parse. ([#2425](https://github.com/paritytech/smoldot/pull/2425))
+- Add support for the `state_getKeys` JSON-RPC method.
 
 ## 0.6.20 - 2022-06-23